extended the functionality for all blocks in section-metadata
